26/7/2017 10:00 Version 1.0. Refactored the projects, see README.txt
